+++ src/dds.ts
@@ -6,7 +6,7 @@ import { getDataFromDB } from 'helpers/db';
 // import { log } from 'helpers/log';
 import { router } from 'router';
 import { legacy } from 'router/legacyRouter';
-import { routerAuth } from 'router/auth';
+import { getUserPermissions } from 'router/admin';
 import { initiateMainUpdate } from 'router/update';
 import config from '../config.json';
 
@@ -67,7 +67,6 @@ app.use(passport.initialize());
 app.use(passport.session());
 
 app.use('/api', router);
-app.use('/auth', routerAuth);
 app.use('/rest', legacy);
 
 app.listen(config.PORT, () =>
@@ -85,3 +84,61 @@ const update = async () => {
 };
 
 setInterval(update, 60000); // TODO change to 600000
+
+// --------------------
+// ------ AUTH --------
+// --------------------
+
+const CLIENT_HOME_PAGE_URL = `${
+  process.env.NODE_ENV === 'development'
+    ? 'http://localhost:3000'
+    : 'http://masochist.me'
+}/home`;
+const CLIENT_ERROR_PAGE_URL = `${
+  process.env.NODE_ENV === 'development'
+    ? 'http://localhost:3000'
+    : 'http://masochist.me'
+}/error`;
+
+const routerAuth = express.Router();
+
+app.use('/auth', routerAuth);
+
+routerAuth.get('/steam', passport.authenticate('steam'));
+
+routerAuth.get('/steam/success', (req: any, res: any): void => {
+  // Successful authentication, redirect home.
+  if (req.user) {
+    const userData = {
+      success: true,
+      message: 'User has successfully authenticated!',
+      user: req.user,
+      cookies: req.cookies,
+    };
+    res.json(userData);
+  } else {
+    res.status(401).json({ error: 'User is not authenticated.' });
+  }
+});
+
+routerAuth.get('/steam/error', (req: any, res: any): void => {
+  res.status(401).json({
+    success: false,
+    message: 'User failed to authenticate!',
+  });
+});
+
+routerAuth.get(
+  '/steam/redirect',
+  passport.authenticate('steam', {
+    successRedirect: CLIENT_HOME_PAGE_URL,
+    failureRedirect: CLIENT_ERROR_PAGE_URL,
+  }),
+);
+
+routerAuth.get('/steam/logout', (req: any, res: any): void => {
+  req.logout();
+  res.redirect(CLIENT_HOME_PAGE_URL);
+});
+
+routerAuth.get('/permissions/:steamid', getUserPermissions);
